Removed IDFT2MAConverter interface - (Task #38)
- Removed uneccessary interface
- restructured packages

---
Task #38: Unify DFT2MaConverter and its interface
