+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/converter/dft2ma/DFT2MAConverter.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.explicit;
+package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -20,8 +20,7 @@ import java.util.Queue;
 import java.util.Set;
 
 import de.dlr.sc.virsat.fdir.core.markov.MarkovAutomaton;
-import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.DFTState;
-import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.IDFT2MAConverter;
+import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.semantics.DFTSemantics;
 import de.dlr.sc.virsat.model.extension.fdir.model.BasicEvent;
 import de.dlr.sc.virsat.model.extension.fdir.model.Fault;
 import de.dlr.sc.virsat.model.extension.fdir.model.FaultTreeNode;
@@ -37,22 +36,27 @@ import de.dlr.sc.virsat.model.extension.fdir.util.FaultTreeHolder;
  *
  */
 
-public class ExplicitDFT2MAConverter implements IDFT2MAConverter {
+public class DFT2MAConverter {
 	private DFTSemantics dftSemantics = DFTSemantics.createNDDFTSemantics();
 	
 	private FaultTreeNode root;
 	
 	private MarkovAutomaton<DFTState> ma;
-	private ExplicitDFTState initial;
+	private DFTState initial;
 
 	private Set<IDFTEvent> events;
 	private Set<FaultTreeNode> transientNodes;
 	private Set<BasicEvent> orderDependentBasicEvents;
-	private Map<Set<BasicEvent>, List<ExplicitDFTState>> mapUnorderedBesToMarkovianDFTStates;
+	private Map<Set<BasicEvent>, List<DFTState>> mapUnorderedBesToMarkovianDFTStates;
 	private FaultTreeHolder ftHolder;
 	private RecoveryStrategy recoveryStrategy;
 	
-	@Override
+	/**
+	 * Converts a fault tree with the passed node as a root to a
+	 * Markov automaton.
+	 * @param root a fault tree node used as a root node for the conversion
+	 * @return the generated Markov automaton resulting from the conversion
+	 */
 	public MarkovAutomaton<DFTState> convert(FaultTreeNode root) {
 		this.root = root;
 		
@@ -144,29 +148,29 @@ public class ExplicitDFT2MAConverter implements IDFT2MAConverter {
 		
 		createInitialState();
 		
-		Queue<ExplicitDFTState> toProcess = new LinkedList<>();
+		Queue<DFTState> toProcess = new LinkedList<>();
 		toProcess.offer(initial);
 		
 		while (!toProcess.isEmpty()) {
-			ExplicitDFTState state = toProcess.poll();
+			DFTState state = toProcess.poll();
 			List<IDFTEvent> occurableEvents = getOccurableEvents(state);
 			
 			for (IDFTEvent event : occurableEvents) {
-				ExplicitDFTState baseSucc = dftSemantics.getStateGenerator().generateState(state);
+				DFTState baseSucc = dftSemantics.getStateGenerator().generateState(state);
 				baseSucc.setRecoveryStrategy(state.getRecoveryStrategy());
 				event.execute(baseSucc, orderDependentBasicEvents, transientNodes);
 				Double rate = event.getRate(state);
 				
-				List<ExplicitDFTState> succs = new ArrayList<>();
+				List<DFTState> succs = new ArrayList<>();
 				succs.add(baseSucc);
 				
-				Map<ExplicitDFTState, List<RecoveryAction>> mapStateToRecoveryActions = new HashMap<>();
+				Map<DFTState, List<RecoveryAction>> mapStateToRecoveryActions = new HashMap<>();
 				mapStateToRecoveryActions.put(baseSucc, new ArrayList<RecoveryAction>());
 				
 				List<FaultTreeNode> changedNodes = dftSemantics.updateFaultTreeNodeToFailedMap(ftHolder, state, 
 						succs, mapStateToRecoveryActions, event);
 				
-				ExplicitDFTState markovSucc = null;
+				DFTState markovSucc = null;
 				if (succs.size() > 1) { 
 					if (recoveryStrategy != null) {
 						Set<FaultTreeNode> occuredEvents = dftSemantics.extractRecoveryActionInput(ftHolder, baseSucc, event, changedNodes);
@@ -180,9 +184,9 @@ public class ExplicitDFT2MAConverter implements IDFT2MAConverter {
 				
 				}
 				
-				for (ExplicitDFTState succ : succs) {
+				for (DFTState succ : succs) {
 					succ.failDontCares(changedNodes, orderDependentBasicEvents);
-					ExplicitDFTState equivalentState = getEquivalentState(succ);
+					DFTState equivalentState = getEquivalentState(succ);
 					
 					if (equivalentState == succ) {
 						ma.addState(succ);
@@ -221,7 +225,7 @@ public class ExplicitDFT2MAConverter implements IDFT2MAConverter {
 	 * @param state the current state
 	 * @return the list of all events that can occur
 	 */
-	private List<IDFTEvent> getOccurableEvents(ExplicitDFTState state) {
+	private List<IDFTEvent> getOccurableEvents(DFTState state) {
 		List<IDFTEvent> occurableEvents = new ArrayList<>();
 		for (IDFTEvent event : events) {
 			if (event.canOccur(state)) {
@@ -236,16 +240,16 @@ public class ExplicitDFT2MAConverter implements IDFT2MAConverter {
 	 * @param state the state for which we want to check if an equivalent one exists
 	 * @return an equivalent state to the passed one or the state itself if not equivalent state exists
 	 */
-	private ExplicitDFTState getEquivalentState(ExplicitDFTState state) {
-		List<ExplicitDFTState> states = mapUnorderedBesToMarkovianDFTStates.get(state.unorderedBes);
+	private DFTState getEquivalentState(DFTState state) {
+		List<DFTState> states = mapUnorderedBesToMarkovianDFTStates.get(state.unorderedBes);
 		if (states == null) {
-			List<ExplicitDFTState> dftStates = new ArrayList<>();
+			List<DFTState> dftStates = new ArrayList<>();
 			dftStates.add(state);
 			mapUnorderedBesToMarkovianDFTStates.put(state.unorderedBes, dftStates);
 			return state;
 		}
 		
-		for (ExplicitDFTState other : states) {
+		for (DFTState other : states) {
 			if (state.isEquivalent(other)) {
 				return other;
 			}
@@ -260,16 +264,22 @@ public class ExplicitDFT2MAConverter implements IDFT2MAConverter {
 	 * Get the initial state of the markov automaton
 	 * @return the initial state of the markov automaton
 	 */
-	public ExplicitDFTState getInitial() {
+	public DFTState getInitial() {
 		return initial;
 	}
-
-	@Override
+	
+	/**
+	 * Sets the node semantics for the converter
+	 * @param dftSemantics the node semantics of the dft nodes
+	 */
 	public void setSemantics(DFTSemantics dftSemantics) {
 		this.dftSemantics = dftSemantics;
 	}
 	
-	@Override
+	/**
+	 * Sets the recovery strategy
+	 * @param recoveryStrategy the recovery strategy
+	 */
 	public void setRecoveryStrategy(RecoveryStrategy recoveryStrategy) {
 		this.recoveryStrategy = recoveryStrategy;
 	}
