+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/converter/dft2ma/IDFTEvent.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.explicit;
+package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma;
 
 import java.util.Set;
 
@@ -26,14 +26,14 @@ public interface IDFTEvent {
 	 * @param state the current state
 	 * @return the occurence rate
 	 */
-	double getRate(ExplicitDFTState state);
+	double getRate(DFTState state);
 	
 	/**
 	 * Checks if a given fault event can occur in a given state.
 	 * @param state the state
 	 * @return true iff the event can occur in the given state
 	 */
-	boolean canOccur(ExplicitDFTState state);
+	boolean canOccur(DFTState state);
 	
 	/**
 	 * Executes a single basic event
@@ -41,7 +41,7 @@ public interface IDFTEvent {
 	 * @param orderDependentBasicEvents set of all basic events that are order dependent 
 	 * @param transientNodes set of all transient nodes
 	 */
-	void execute(ExplicitDFTState state, Set<BasicEvent> orderDependentBasicEvents, Set<FaultTreeNode> transientNodes);
+	void execute(DFTState state, Set<BasicEvent> orderDependentBasicEvents, Set<FaultTreeNode> transientNodes);
 	
 	/**
 	 * Gets the related DFT node
