+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/converter/dft2ma/TimeEvent.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.explicit;
+package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma;
 
 import java.util.Set;
 
@@ -37,18 +37,18 @@ public class TimeEvent implements IDFTEvent {
 	}
 	
 	@Override
-	public double getRate(ExplicitDFTState state) {
+	public double getRate(DFTState state) {
 		return 1 / time;
 	}
 
 	@Override
-	public boolean canOccur(ExplicitDFTState state) {
+	public boolean canOccur(DFTState state) {
 		RecoveryStrategy raStrategy = (RecoveryStrategy) state.getRecoveryStrategy();
 		return raStrategy.getCurrentState().equals(raState);
 	}
 
 	@Override
-	public void execute(ExplicitDFTState state, Set<BasicEvent> orderDependentBasicEvents,
+	public void execute(DFTState state, Set<BasicEvent> orderDependentBasicEvents,
 			Set<FaultTreeNode> transientNodes) {
 		state.setRecoveryStrategy(state.getRecoveryStrategy().onTime(time));
 	}
