--- de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/converter/dft2ma/explicit/DelaySemantics.java
@@ -1,39 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2008-2019 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *******************************************************************************/
-package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.explicit;
-
-import java.util.List;
-
-import de.dlr.sc.virsat.model.extension.fdir.model.FaultTreeNode;
-import de.dlr.sc.virsat.model.extension.fdir.util.FaultTreeHolder;
-
-/**
- * Simple propagation semantics. A node with this semantics fails iff a child fails.
- * @author muel_s8
- *
- */
-
-public class DelaySemantics implements INodeSemantics {
-
-	@Override
-	public boolean handleUpdate(FaultTreeNode node, ExplicitDFTState state, ExplicitDFTState pred,
-			FaultTreeHolder ftHolder, GenerationResult generationResult) {
-		List<FaultTreeNode> children = ftHolder.getMapNodeToChildren().get(node);
-		
-		for (FaultTreeNode child : children) {
-			if (state.hasFaultTreeNodeFailed(child)) {
-				return state.setFaultTreeNodeFailing(node, true);
-			}
-		}
-		
-		return state.setFaultTreeNodeFailing(node, false) | state.setFaultTreeNodeFailed(node, false);
-	}
-
-}
