+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/converter/dft2ma/semantics/StandardSPARESemantics.java
@@ -7,10 +7,12 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.explicit;
+package de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.semantics;
 
 import java.util.List;
 
+import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.DFTState;
+import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.GenerationResult;
 import de.dlr.sc.virsat.model.extension.fdir.model.FaultTreeNode;
 import de.dlr.sc.virsat.model.extension.fdir.model.SPARE;
 import de.dlr.sc.virsat.model.extension.fdir.util.FaultTreeHolder;
@@ -25,7 +27,7 @@ import de.dlr.sc.virsat.model.extension.fdir.util.FaultTreeHolder;
 public class StandardSPARESemantics implements INodeSemantics {
 	
 	@Override
-	public boolean handleUpdate(FaultTreeNode node, ExplicitDFTState state, ExplicitDFTState pred, FaultTreeHolder ftHolder,
+	public boolean handleUpdate(FaultTreeNode node, DFTState state, DFTState pred, FaultTreeHolder ftHolder,
 			GenerationResult generationResult) {
 		List<FaultTreeNode> spares = ftHolder.getMapNodeToSpares().get(node);
 		List<FaultTreeNode> children = ftHolder.getMapNodeToChildren().get(node);
@@ -95,7 +97,7 @@ public class StandardSPARESemantics implements INodeSemantics {
 	 * @param generationResult accumulator for state space generation results
 	 * @return constant true
 	 */
-	protected boolean performClaim(SPARE node, FaultTreeNode spare, ExplicitDFTState state, 
+	protected boolean performClaim(SPARE node, FaultTreeNode spare, DFTState state, 
 			GenerationResult generationResult) {
 		state.getSpareClaims().put(spare, node);
 		state.activateNode(spare);
